g6R: Graph Visualisation Engine Widget for R and 'shiny' Apps

Create stunning network experiences powered by the G6 graph visualisation engine JavaScript library https://g6.antv.antgroup.com/en. In 'shiny' mode, modify your graph directly from the server function to dynamically interact nodes and edges. Select you favorite layout among 20 choices. 15 behaviors are available such as interactive edge creation, collapse-expand and brush select. 17 plugins designed to improve the user experience such as a minimap, toolbars and grid lines.

See also

Author

Maintainer: David Granjon dgranjon@ymail.com

Authors:

David Schoch david@schochastics.net

Other contributors:

cynkra GmbH [funder]

Bristol Myers Squibb [funder]