Check if an object is a list of G6 elements

Check if an object is a list of G6 elements

Usage

is_g6_nodes(x)

is_g6_edges(x)

is_g6_combos(x)

Arguments

x

An object of class g6_nodes, g6_edges, or g6_combos.